+++ raster/src/main/scala/geotrellis/raster/merge/MultibandTileMergeMethods.scala
@@ -26,18 +26,14 @@ import geotrellis.vector.Extent
   */
 trait MultibandTileMergeMethods extends TileMergeMethods[MultibandTile] {
 
-  /**
-    * Merge the respective bands of this MultibandTile and the other
-    * one.
-    */
-  def merge(other: MultibandTile): MultibandTile = {
+  def merge(other: MultibandTile, baseCol: Int, baseRow: Int): MultibandTile = {
     val bands: Seq[Tile] =
       for {
         bandIndex <- 0 until self.bandCount
       } yield {
         val thisBand = self.band(bandIndex)
         val thatBand = other.band(bandIndex)
-        thisBand.merge(thatBand)
+        thisBand.merge(thatBand, baseCol, baseRow)
       }
 
     ArrayMultibandTile(bands)